move repository implementations to the agent package
