+++ dive.go
@@ -2,6 +2,7 @@ package dive
 
 import (
 	"context"
+	"fmt"
 	"time"
 
 	"github.com/diveagents/dive/llm"
@@ -187,4 +188,62 @@ type TaskResult struct {
 
 	// Usage tracks LLM token usage
 	Usage llm.Usage
+}
+
+// ListDocumentInput specifies search criteria for documents in a document store
+type ListDocumentInput struct {
+	PathPrefix string
+	Recursive  bool
+}
+
+// ListDocumentOutput is the output for listing documents
+type ListDocumentOutput struct {
+	Items []Document
+}
+
+// DocumentRepository provides read/write access to a set of documents. This could be
+// backed by a local file system, a remote API, or a database.
+type DocumentRepository interface {
+
+	// GetDocument returns a document by name
+	GetDocument(ctx context.Context, name string) (Document, error)
+
+	// ListDocuments lists documents
+	ListDocuments(ctx context.Context, input *ListDocumentInput) (*ListDocumentOutput, error)
+
+	// PutDocument puts a document
+	PutDocument(ctx context.Context, doc Document) error
+
+	// DeleteDocument deletes a document
+	DeleteDocument(ctx context.Context, doc Document) error
+
+	// Exists checks if a document exists by name
+	Exists(ctx context.Context, name string) (bool, error)
+
+	// RegisterDocument assigns a name to a document path
+	RegisterDocument(ctx context.Context, name, path string) error
+}
+
+var ErrThreadNotFound = fmt.Errorf("thread not found")
+
+// Thread represents a chat thread
+type Thread struct {
+	ID        string         `json:"id"`
+	UserID    string         `json:"user_id"`
+	CreatedAt time.Time      `json:"created_at"`
+	UpdatedAt time.Time      `json:"updated_at"`
+	Messages  []*llm.Message `json:"messages"`
+}
+
+// ThreadRepository is an interface for storing and retrieving chat threads
+type ThreadRepository interface {
+
+	// PutThread creates or updates a thread
+	PutThread(ctx context.Context, thread *Thread) error
+
+	// GetThread retrieves a thread by ID
+	GetThread(ctx context.Context, id string) (*Thread, error)
+
+	// DeleteThread deletes a thread by ID
+	DeleteThread(ctx context.Context, id string) error
 }
